k8s-automation-server/requirements.txt
fastapi[all]        
kubernetes           
pyyaml               